refactor(mga): split classify_lineup + classify_frames into own modules; clear classifier_service allowlist
Extracts the two remaining ~300-LOC Claude code paths from
classifier_service.py into sibling modules:

  - single_image_classifier.py — _OUTPUT_SCHEMA_DOC + classify_lineup
    (the re-classify path: one stored screenshot, cannot decide is_lineup)
  - grid_classifier.py — _GRID_OUTPUT_SCHEMA_DOC +
    classify_frames_for_lineup_decision (Strategy A ingest-time path: N
    candidate frames, decides is_lineup + picks STAND/AIM frames)

classifier_service.py is now 267 LOC (was 2029 at session start, 1025
after PR R1). Only shared blocks remain:
  - _GAME_VISUAL_CUES, _GAME_FIRST_RULE
  - _build_reference_text, _check_game_map_consistency,
    _fetch_screenshot_bytes
  - _strip_json_fences, _validate_aim_coord, _validate_grid_index
  - re-export footer for all four extracted classifiers

Both allowlist entries for classifier_service.py REMOVED — it's now well
under both the 1000-LOC hard limit AND the 500-LOC growth-guard threshold.
The only remaining allowlist entry is pipeline.rs (1120 LOC, Tauri CV
orchestrator).

Re-export contract: all four extracted classifiers (single_image,
grid, throw_timing, throw_technique) are re-exported from
classifier_service so existing callers need NO import changes. Confirmed
unchanged for: ingestion_orchestrator, lineup_service, technique_extractor,
throw_localizer, micro_clip_generator, landing_clip_generator, etc.

Tests:
  - TestClassifyLineup + TestStructuredFailureSurfacing patches now target
    single_image_classifier.{settings,anthropic.Anthropic,_fetch_screenshot_bytes}
  - TestClassifyFramesForLineupDecision + TestGridMaxTokens patches target
    grid_classifier.{settings,anthropic.Anthropic}
  - TestGameFirstPromptWiring helpers import _OUTPUT_SCHEMA_DOC from
    single_image_classifier and _GRID_OUTPUT_SCHEMA_DOC from grid_classifier
  - test_grid_schema_format_does_not_break references gc._GRID_OUTPUT_SCHEMA_DOC
  - Full backend suite (676 tests) green
